SecondOrderPerturbationTheory

Documentation for SecondOrderPerturbationTheory.

SecondOrderPerturbationTheory.BinaryConfigure

SecondOrderPerturbationTheory.Coefficience

SecondOrderPerturbationTheory.PickState

SecondOrderPerturbationTheory.ProjectState

SecondOrderPerturbationTheory.ProjectState

SecondOrderPerturbationTheory.ProjectStateBond

SecondOrderPerturbationTheory.ProjectStateBond

SecondOrderPerturbationTheory.SOPT

SecondOrderPerturbationTheory.SOPT

SecondOrderPerturbationTheory.SOPTMatrix

SecondOrderPerturbationTheory.SecondOrderPerturbation

Base.:<<

Base.:<<

QuantumLattices.:⊕

QuantumLattices.:⊗

QuantumLattices.QuantumOperators.matrix

QuantumLattices.QuantumOperators.matrix

QuantumLattices.QuantumOperators.matrix

QuantumLattices.dimension

QuantumLattices.dimension

QuantumLattices.expand

SecondOrderPerturbationTheory.:⊠

SecondOrderPerturbationTheory.coefficience_project

SecondOrderPerturbationTheory.coefficience_project

SecondOrderPerturbationTheory.hamiltonianeff

SecondOrderPerturbationTheory.high_configure

SecondOrderPerturbationTheory.observables_project

SecondOrderPerturbationTheory.projectstate_points

SecondOrderPerturbationTheory.projectstate_points

Getting Started

Examples of second order perturbation theory

Manuals

SecondOrderPerturbationTheory.BinaryConfigure — Type

BinaryConfigure{I<:TargetSpace, P<:Int} <: CompositeDict{P, I}
BinaryConfigure(ps::Pair...)
BinaryConfigure(kv)

Construct BinaryConfigure at a lattice. The local binary configure is given by BinaryConfigure.

SecondOrderPerturbationTheory.Coefficience — Type

Coefficience{P<:Int, I<:AbstractVector{<:Matrix{<:Number}}} <: Action
Coefficience(ob::AbstractVector{<:Matrix{<:Number}}, lattice::AbstractLattice; order::Int=-1, η::Float64=1e-12)
Coefficience(lattice::AbstractLattice, hilbert::Hilbert, terms::Tuple{Vararg{Term}}, p₀::Dict{<:Int, <:ProjectState}; η::Float64=1e-12, order::Int=-1)
Coefficience(observables::Dict{<:Int, <:AbstractVector{<:Matrix{<:Number}}}; η::Float64=1e-12, order::Int= -1)

The observables attribute is a Dict{<:Int, <: AbstractVector{<:Matrix{<:Number}}}. η attribute is truncation of completeness of physical quantities. order attribute choose the order of matrix of effective hamiltonian to obtain the exchange coefficiences.

SecondOrderPerturbationTheory.PickState — Type

PickState{I<:Vector{Vector{Int}}, P<:Int} <: CompositeDict{P, I}
PickState(ps::Pair...)
PickState(kv)

Construct PickState. Pick low-energy states at a lattice.

SecondOrderPerturbationTheory.ProjectState — Type

ProjectState{V<:Real, C<:Number, P<:TargetSpace}

Projected states contain eigenvalues, eigenvectors, and basis of H₀.

SecondOrderPerturbationTheory.ProjectState — Method

ProjectState(ops::Operators, braket::BinaryBases, table; pick::Union{UnitRange{Int}, Vector{Int}, Colon}=:)
ProjectState(ops::Operators, ts::TargetSpace, table, pick::Vector{Vector{Int}})

Construct ProjectState. The pick::Union{UnitRange{Int}, Vector{Int}, Colon} argument picks the low-energy states. The i-th element of arguement pick vector is the loaction of low-energy states in the i-th Sector of TargetSpace.

SecondOrderPerturbationTheory.ProjectStateBond — Type

ProjectStateBond(left::ProjectState, right::ProjectState)

Projected states on a bond. Construct ProjectStateBond by the two ProjectStates defined on point.

SecondOrderPerturbationTheory.ProjectStateBond — Method

ProjectStateBond(left::ProjectState, right::ProjectState, nleft::Int, nright::Int)

The nleft and nright are the number of bits of left bit shift operator for left and right ProjectState, respectively.

SecondOrderPerturbationTheory.SOPT — Type

SOPT{L<:AbstractLattice, G₁<:OperatorGenerator, G₀<:OperatorGenerator, PT<:SecondOrderPerturbation} <: Frontend

Second order perturbation theory method of a electronic quantum lattice system.

SecondOrderPerturbationTheory.SOPT — Method

SOPT(lattice::AbstractLattice, hilbert::Hilbert, terms₁::Tuple{Vararg{Term}}, terms₀::Tuple{Vararg{Term}}, binaryconfigure::BinaryConfigure, lowstate::PickState; neighbors::Union{Nothing, Int, Neighbors}=nothing, boundary::Boundary=plain)

Construct the second order perturbation method for a quantum lattice system.

SecondOrderPerturbationTheory.SOPTMatrix — Type

SOPTMatrix(bond::Bond, P₀::ProjectStateBond, m₀::Matrix, m₂::Matrix)

Matrix representation of the low-energy hamiltionian. The order of basis of representation is the order of (space of bond[2], space of bond[1]), i.e. (left space of bond[1])⊗(right space of bond[2])

Arguments

-bond: bond of lattice -P₀: projected state -m₀: matrix representation of the zeroth order of the low-energy hamiltionian -m₂: matrix representation of second order of the low-energy hamiltionian

SecondOrderPerturbationTheory.SecondOrderPerturbation — Type

SecondOrderPerturbation{B<:BinaryConfigure, L<:PickState} <: Transformation
(::SecondOrderPerturbation)(H₁::OperatorGenerator, p₀::Dict{T,<:ProjectState}, qₚ::Dict{T,<:ProjectState}, qₘ::Dict{T,<:ProjectState}, bond::Bond) where T<:Int  -> SOPTMatrix
(::SecondOrderPerturbation)(H₀::OperatorGenerator, H₁::OperatorGenerator, bond::Bond) -> SOPTMatrix

A type.

Base.:<< — Method

Base.:(<<)(bs::BinaryBases, n::Int) -> BinaryBases

Base.:<< — Method

Base.:(<<)(ps::ProjectState, n::Int) -> ProjectState

Left bit shift opeartor. The BinaryBasis is left shifted by n bits.

QuantumLattices.:⊕ — Method

⊕(ps₁::ProjectState, ps₂::ProjectState) -> ProjectState

Get the direct sum of two projected states.

QuantumLattices.:⊗ — Method

⊗(ps₁::ProjectState, ps₂::ProjectState) -> ProjectState

Get the direct product of two sets of projected states.

QuantumLattices.QuantumOperators.matrix — Method

matrix(ops::Operators, ts₁::TargetSpace, ts₂::TargetSpace, table) -> Matrix
matrix(ops::Operators, ts::TargetSpace, table) -> Matrix

Get the matrix of direct sum of submatrices.

QuantumLattices.QuantumOperators.matrix — Method

matrix(sopt::SOPT, bond::Bond) -> SOPTMatrix

QuantumLattices.QuantumOperators.matrix — Method

matrix(sopt::SOPT) -> Vector

Obtain SOPTMatrix on all bonds.

QuantumLattices.dimension — Method

dimension(ps::ProjectState) -> Int

The dimension of local low-energy hilbert space.

QuantumLattices.dimension — Method

dimension(ps::ProjectState,::Type{T}) where T<:TargetSpace -> Int

The dimension of target space basis.

QuantumLattices.expand — Method

expand(gen::OperatorGenerator, bond::Bond) -> Operators

SecondOrderPerturbationTheory.:⊠ — Method

⊠(bs₁::BinaryBases, bs₂::BinaryBases) -> Tuple{BinaryBases, Vector{Int}}

Get the direct product of two sets of binary bases, and the permutation vector.

SecondOrderPerturbationTheory.coefficience_project — Method

coefficience_project(st::SOPT, cof::Coefficience; η::Float64=1e-14) -> OperatorGenerator

Only support the pseudospin-1/2 case. The zeeman term is ommited when the ground states are not Krameter states.

SecondOrderPerturbationTheory.coefficience_project — Method

coefficience_project(m₂::Matrix{<:Number}, gsg::AbstractVector{T}, nshape::Tuple{Int,Int}; η::Float64=1e-12) where T<:Matrix{<:Number} -> Matrix
coefficience_project(m₂::Matrix{<:Number}, bond::Bond, coeff::Coefficience) -> Matrix
coefficience_project(soptm::SOPTMatrix, coeff::Coefficience) -> Matrix

Get the coefficience of exchange interaction. The row index corresponds to bond[1], column index corresponds to bond[2].

SecondOrderPerturbationTheory.hamiltonianeff — Method

hamiltonianeff(psp::ProjectState, psq::ProjectState, h1::Operators, table::Table)  ->Tuple{Matrix, Matrix}

Get the effective Hamiltonian, the first and second terms of the result correspond to the zero-th and 2nd perturbations respectively.

SecondOrderPerturbationTheory.high_configure — Method

high_configure(bc::BinaryConfigure, ls::PickState) -> Tuple
high_configure(bc::BinaryConfigure, ls::PickState, pids::AbstractVector{Int}) -> Tuple

Get the high-energy configure of local space.

SecondOrderPerturbationTheory.observables_project — Method

observables_project(terms::Tuple{Vararg{Term}}, point::Point, hilbert::Hilbert, psp::ProjectState) -> Vector

The order of Term in terms (Onsite) tuple determines the order of matrix in the result.

SecondOrderPerturbationTheory.projectstate_points — Method

projectstate_points(bc::BinaryConfigure, ls::PickState, H₀::OperatorGenerator) -> Tuple{Dict{Int, ProjectState}, Dict{Int, ProjectState}, Dict{Int, ProjectState}}
projectstate_points(bc::BinaryConfigure, ls::PickState, H₀::OperatorGenerator, points::AbstractVector{<:Point}) -> Tuple{Dict{Int, ProjectState}, Dict{Int, ProjectState}, Dict{Int, ProjectState}}

Construct the ProjectState` type of low-energy states within N-particle space, high-energy states with (N+1)-particle space, and high-energy states with (N-1)-particle space.

SecondOrderPerturbationTheory.projectstate_points — Method

projectstate_points(sopt::SOPT) -> Tuple{Dict{Int, ProjectState}, Dict{Int, ProjectState}, Dict{Int, ProjectState}}

Construct ProjectState on all points.