SecondOrderPerturbationTheory

Documentation for SecondOrderPerturbationTheory.

SecondOrderPerturbationTheory.BinaryConfigure

SecondOrderPerturbationTheory.Coefficience

SecondOrderPerturbationTheory.Coefficience

SecondOrderPerturbationTheory.PickState

SecondOrderPerturbationTheory.ProjectState

SecondOrderPerturbationTheory.ProjectState

SecondOrderPerturbationTheory.ProjectStateBond

SecondOrderPerturbationTheory.ProjectStateBond

SecondOrderPerturbationTheory.SOPT

SecondOrderPerturbationTheory.SOPT

SecondOrderPerturbationTheory.SOPTMatrix

SecondOrderPerturbationTheory.SecondOrderPerturationMetric

SecondOrderPerturbationTheory.SecondOrderPerturbation

Base.:<<

Base.:<<

QuantumLattices.:⊕

QuantumLattices.:⊗

QuantumLattices.QuantumOperators.matrix

QuantumLattices.QuantumOperators.matrix

QuantumLattices.dimension

QuantumLattices.dimension

QuantumLattices.expand

SecondOrderPerturbationTheory.:⊠

SecondOrderPerturbationTheory.SpinOperatorGenerator

SecondOrderPerturbationTheory.coefficience_project

SecondOrderPerturbationTheory.coefficience_project

SecondOrderPerturbationTheory.hamiltonianeff

SecondOrderPerturbationTheory.observables_project

SecondOrderPerturbationTheory.projectstate_points

SecondOrderPerturbationTheory.projectstate_points

Getting Started

Examples of second order perturbation theory

Manuals

SecondOrderPerturbationTheory.BinaryConfigure — Type

BinaryConfigure{I<:TargetSpace, P<:Int} <: CompositeDict{P, I}
BinaryConfigure(ps::Pair...)
BinaryConfigure(kv)

Construct BinaryConfigure at a lattice. The local binary configure is given by BinaryConfigure.

SecondOrderPerturbationTheory.Coefficience — Type

Coefficience{P<:Int, I<:AbstractVector{<:Matrix{<:Number}}} <: Action

Decompose the projected-hamiltionian matrix with respect to input physical observables.

SecondOrderPerturbationTheory.Coefficience — Method

Coefficience(bonds::AbstractVector{<:Bond}, observables::Dict{Int, I}; options...) where {I<:Union{AbstractVector{<:AbstractMatrix}, Tuple{Vararg{Term}}}}
Coefficience(bonds::AbstractVector{<:Bond}, observables::Dict{Int, I}; options...) where {I<:Union{AbstractVector{<:AbstractMatrix}, Tuple{Vararg{Term}}}}
Coefficience(bonds::AbstractVector{<:Bond}, npoints::Int, ob::I=Tuple{}(); options...) where {I<:Union{AbstractVector{<:AbstractMatrix}, Tuple{Vararg{Term}}}}

η attribute in options is truncation of completeness of physical quantities. order attribute (order=-1,0,2) choose the order of matrix of effective hamiltonian to obtain the exchange coefficiences. npoints = length(lattice).

SecondOrderPerturbationTheory.PickState — Type

PickState{I<:Vector{Vector{Int}}, P<:Int} <: CompositeDict{P, I}
PickState(ps::Pair...)
PickState(kv)

Construct PickState. Pick low-energy states at a lattice.

SecondOrderPerturbationTheory.ProjectState — Type

ProjectState{V<:Real, C<:Number, P<:TargetSpace}

Projected states contain eigenvalues, eigenvectors, and basis of H₀.

SecondOrderPerturbationTheory.ProjectState — Method

ProjectState(ops::Operators, braket::BinaryBases, table; pick::Union{UnitRange{Int}, Vector{Int}, Colon}=:)
ProjectState(ops::Operators, ts::TargetSpace, table, pick::Vector{Vector{Int}})

Construct ProjectState. The pick::Union{UnitRange{Int}, Vector{Int}, Colon} argument picks the low-energy states. The i-th element of arguement pick vector is the loaction of low-energy states in the i-th Sector of TargetSpace.

SecondOrderPerturbationTheory.ProjectStateBond — Type

ProjectStateBond

A type. Projected states on a bond. Construct ProjectStateBond by the ProjectStates defined on points. When bond.kind==2, the left (right) corresponds to the first (second) point in bond::Bond.

SecondOrderPerturbationTheory.ProjectStateBond — Method

ProjectStateBond(left::ProjectState, right::ProjectState, nleft::Int, nright::Int)

The nleft and nright are the number of bits of left bit shift operator for left and right ProjectState, respectively.

SecondOrderPerturbationTheory.SOPT — Type

SOPT{L<:AbstractLattice, G₁<:OperatorGenerator, G₀<:OperatorGenerator, PT<:SecondOrderPerturbation} <: Frontend

Second order perturbation theory method of a electronic quantum lattice system.

SecondOrderPerturbationTheory.SOPT — Method

SOPT(lattice::AbstractLattice, hilbert::Hilbert, terms₁::Tuple{Vararg{Term}}, terms₀::Tuple{Vararg{Term}}, binaryconfigure::BinaryConfigure, lowstate::PickState; neighbors::Union{Nothing, Int, Neighbors}=nothing, boundary::Boundary=plain)

Construct the second order perturbation method for a quantum lattice system.

SecondOrderPerturbationTheory.SOPTMatrix — Type

SOPTMatrix(bond::Bond, P₀::ProjectStateBond, m₀::Matrix, m₂::Matrix)

Matrix representation of the low-energy hamiltionian. The proper order for representing basis vectors is to start with the left space (bond[1]) and then proceed to the right space (bond[2]), i.e. kron(right, left).

Arguments

-bond: bond of lattice -P₀: projected state -m₀: matrix representation of the zeroth order (<H₀>) of the low-energy hamiltionian -m₀₁: matrix representation of the zeroth order (<H₁>) of the low-energy hamiltionian -m₂: matrix representation of second order of the low-energy hamiltionian

SecondOrderPerturbationTheory.SecondOrderPerturationMetric — Type

SecondOrderPerturationMetric <: Metric

Define the table metric, (:spin, :orbital) for a point, (:rcoordinate,:site, :spin, :orbital) for a bond.

SecondOrderPerturbationTheory.SecondOrderPerturbation — Type

SecondOrderPerturbation{B<:BinaryConfigure, L<:PickState} <: Transformation
(::SecondOrderPerturbation)(H₁::OperatorGenerator, p₀::Dict{T,<:ProjectState}, qₚ::Dict{T,<:ProjectState}, qₘ::Dict{T,<:ProjectState}, bond::Bond) where T<:Int  -> SOPTMatrix
(::SecondOrderPerturbation)(H₀::OperatorGenerator, H₁::OperatorGenerator, bond::Bond) -> SOPTMatrix

A type.

Base.:<< — Method

Base.:(<<)(bs::BinaryBases, n::Int) -> BinaryBases

Base.:<< — Method

Base.:(<<)(ps::ProjectState, n::Int) -> ProjectState

Left bit shift opeartor. The BinaryBasis is left shifted by n bits.

QuantumLattices.:⊕ — Method

⊕(ps₁::ProjectState, ps₂::ProjectState) -> ProjectState

Get the direct sum of two projected states.

QuantumLattices.:⊗ — Method

⊗(ps₁::ProjectState, ps₂::ProjectState) -> ProjectState

Get the direct product of two sets of projected states.

QuantumLattices.QuantumOperators.matrix — Method

matrix(ops::Operators, ts₁::TargetSpace, ts₂::TargetSpace, table) -> Matrix
matrix(ops::Operators, ts::TargetSpace, table) -> Matrix

Get the matrix of direct sum of submatrices.

QuantumLattices.QuantumOperators.matrix — Method

matrix(sopt::SOPT, bond::Bond) -> SOPTMatrix

Return matrix of the projected hamiltionian onto a bond of which the length is greater than 1.

QuantumLattices.dimension — Method

dimension(ps::ProjectState) -> Int

The dimension of local low-energy hilbert space.

QuantumLattices.dimension — Method

dimension(ps::ProjectState,::Type{T}) where T<:TargetSpace -> Int

The dimension of target space basis.

QuantumLattices.expand — Method

expand(gen::OperatorGenerator, bond::Bond) -> Operators

SecondOrderPerturbationTheory.:⊠ — Method

⊠(bs₁::BinaryBases, bs₂::BinaryBases) -> Tuple{BinaryBases, Vector{Int}}

Get the direct product of two sets of binary bases, and the permutation vector. oxtimes=⊠

SecondOrderPerturbationTheory.SpinOperatorGenerator — Method

SpinOperatorGenerator(st::SOPT, coef::Coefficience; η::Float64=1e-14) -> OperatorGenerator

Only support the pseudospin-1/2 case. The zeeman term (Onsite term) is ommited when the ground states are not Krameter states.

SecondOrderPerturbationTheory.coefficience_project — Function

coefficience_project(m₂::Matrix{<:Number}, bond::Bond, ob::Dict{Int, <:AbstractVector{<:Matrix{<:Number}}}, η::Float64=1e-14) -> Matrix

Get the coefficience of exchange interaction. The row index corresponds to bond[1], column index corresponds to bond[2].

SecondOrderPerturbationTheory.coefficience_project — Method

coefficience_project(matbd::SOPTMatrix, coef::Coefficience, hilbert::Hilbert) -> Matrix

SecondOrderPerturbationTheory.hamiltonianeff — Function

hamiltonianeff(psp::ProjectStateBond, psq::ProjectStateBond, h1::Operators, table::Table, bond::Union{Bond, Nothing}=nothing)   ->Tuple{Matrix, Matrix, Matrix}

Get the effective Hamiltonian, the first and second terms of the result correspond to the zero-th and 2nd perturbations respectively.

SecondOrderPerturbationTheory.observables_project — Method

observables_project(terms::Tuple{Vararg{Term}}, point::Point, hilbert::Hilbert, psp::ProjectState) -> Vector{Matrix}

The order of Term in terms (Onsite) tuple determines the order of matrix in the result.

SecondOrderPerturbationTheory.projectstate_points — Method

projectstate_points(bc::BinaryConfigure, ls::PickState, H₀::OperatorGenerator, points::AbstractVector{<:Point}) -> Tuple{Dict, Dict, Dict, Dict}

Construct the ProjectState` type of low-energy states with N-particle space, high-energy states with (N+1)-particle space, high-energy states with (N-1)-particle space, and high-energy states with N-particle space.

SecondOrderPerturbationTheory.projectstate_points — Method

projectstate_points(sopt::SOPT) -> Tuple{Dict{Int, ProjectState}, Dict{Int, ProjectState}, Dict{Int, ProjectState}, Dict{Int, ProjectState}}

Construct ProjectState on all points. Construct the ProjectState` type of low-energy states with N-particle space, high-energy states with (N+1)-particle space, high-energy states with (N-1)-particle space, and high-energy states with N-particle space.